refactor: implement smooth tab cross-fade animations, add radio icon for shared tabs, and enforce integer constraints on config schemas

+++ src/schemas/configSchema.ts
@@ -25,10 +25,12 @@ import { z } from "zod";
 export const ConfigSchema = z
   .object({
     theme: z.enum(["aterm-dark", "aterm-light", "nord"]).default("aterm-dark"),
-    fontSize: z.number().min(8).max(32).default(12),
+    // fontSize must be an integer to match Rust `u8` (src-tauri/src/config.rs:default_font_size);
+    // floats like 22.5 would pass `z.number().min().max()` but fail serde `u8` deserialization.
+    fontSize: z.number().int().min(8).max(32).default(12),
     shell: z.string().default(""),
     fontFamily: z.string().default("JetBrains Mono, monospace"),
-    scrollback: z.number().min(100).max(10000).default(1500),
+    scrollback: z.number().int().min(100).max(10000).default(1500),
   })
   .passthrough();
